// console/src/generated/RuntimeSettingsDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LogFormatDto } from "./LogFormatDto";
import type { LogLevelDto } from "./LogLevelDto";

export type RuntimeSettingsDto = { proxy: string | null, inherit_system_proxy: boolean, file_upload_max_in_flight: number, cors_origins: Array<string>, trusted_proxies: Array<string>, max_attempts: number, max_in_flight: number, log_level: LogLevelDto, log_format: LogFormatDto, };